[Validation] perf improvements for fragment cycle detection
This converts the O(N^2) cycle detector to an O(N) detector and optimizes the visit to short-circuit where parts of the validation visitor were unnecessary.

On an internal very large stress test query this was previously ~30s and after this diff runs in 15ms.
Commit:
4cf2190b54fefc34a7b37d22a489933fc15e14ce [4cf2190]
Parents:
a738c6d90d
Author:
Lee Byron <lee@leebyron.com>
Date:
13 November 2015 at 3:47:01 PM SGT

[Validation] Performance improvements

Fairly dramatic improvement of some validation rules by short-circuiting branches which do not need to be checked and memoizing the process of collecting fragment spreads from a given context - which is necessary by multiple rules.
Commit:
0bc9088187b9902ab19c0ec34e0e9f036dc9d9ea [0bc9088]
Parents:
7278b7c92c
Author:
Lee Byron <lee@leebyron.com>
Date:
13 November 2015 at 4:26:05 PM SGT
Labels:
HEAD

[Validation] Report errors rather than return them

This replaces the mechanism of returning errors or lists of errors from a validator step to instead report those errors via calling a function on the context.

This simplifies the implementation of the visitor mechanism, but also opens the doors for future rules being specified as warnings instead of errors.