[Validation] Factor out and memoize recursively referenced fragments.
This adds a new method to the validation context which when given an Operation definition, returns a list of all Fragment definitions recursively referenced via fragment spreads.

This new method is then used to ensure no fragments are unused.

The implementation of this method is the same in principle as the one which used to be inline in the validation rule, but has been unfolded from recursion to use a while loop.
Commit:
eef8d97f64b5b5fa0df79435c4fe237976867573 [eef8d97]
Parents:
568dc52a4f
Author:
Lee Byron <lee@leebyron.com>
Date:
17 November 2015 at 9:31:46 AM SGT