[Validation] Report errors rather than return them
This replaces the mechanism of returning errors or lists of errors from a validator step to instead report those errors via calling a function on the context.

This simplifies the implementation of the visitor mechanism, but also opens the doors for future rules being specified as warnings instead of errors.

Commit:
5e545cce0104708a4ac6e994dd5f837d1d30a09b [5e545cc]
Parents:
ef7c755c58
Author:
Lee Byron <lee@leebyron.com>
Date:
17 November 2015 at 11:32:13 AM SGT