Add GraphQLSchema types field
This is a rebased and updated version of @tgriesser's #199. I further extended it to completely remove the side-effectful mutation of Interface types rather than just deferring that mutation to schema creation time.

This introduces a *breaking* change to the Type System API. Now, any individual Interface type does not have the required information to answer `getPossibleTypes` or `isPossibleType` without knowing the other types in the Schema. These methods were moved to the Schema API, accepting the abstract type as the first parameter.

This also introduces a *breaking* change to the type comparator functions: `isTypeSubTypeOf` and `doTypesOverlap` which now require a Schema as a first argument.

Commit:
6a1f23e1f9c1e6bf4cea837bc9bb6eae0fb5c382 [6a1f23e]
Parents:
a781b556ee
Author:
Lee Byron <lee@leebyron.com>
Date:
23 March 2016 at 1:17:43 PM SGT
Commit Date:
25 March 2016 at 6:35:39 AM SGT